ml-service/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
numpy